`TSGetWorkspaceSymbolsFunc` + `TSSearch`
`TSGetWorkspaceSymbolsFunc` now creates a location list parsable output
instead of return a raw result from tsserver.
It seems like this function was used in an attempt to a denite.vim input
but it was not working properly. I did not debug it :/ (not a denite
user).

Renamed instances of "Workplace" to "Workspace" as I think this was a
typo.

Removed commented `TSGetDocSymbols` function as an uncommented exact version
exists in the same file.