Fix errors on Windows platform
This patch implements a fix for mhartington/nvim-typescript#119 as
proposed by @cmcclellen.

The path to the tsserver command is run through `os.path.normpath()` to
ensure it uses backslashes on Windows, and `shell=True` is passed to
appropriate subprocess calls so that `.cmd` files are consideres when
invoking the command.